FEAT: Create add fab
